third_party/blink/web_tests/inspector-protocol/page/enable-file-chooser-opened-event-expected.txt

Tests that Page.enable(enableFileChooserOpenedEvent) enables `Page.fileChooserOpened` events

Running test: testClickFileInput
Page.fileChooserOpened {
    method : Page.fileChooserOpened
    params : {
        backendNodeId : <number>
        frameId : <string>
        mode : selectSingle
    }
    sessionId : <string>
}

Running test: testShowPickerAPI
Page.fileChooserOpened {
    method : Page.fileChooserOpened
    params : {
        backendNodeId : <number>
        frameId : <string>
        mode : selectSingle
    }
    sessionId : <string>
}

Running test: testOpenFilePickerAPI
Page.fileChooserOpened {
    method : Page.fileChooserOpened
    params : {
        frameId : <string>
        mode : selectSingle
    }
    sessionId : <string>
}

Running test: testSaveFilePickerAPI
Page.fileChooserOpened {
    method : Page.fileChooserOpened
    params : {
        frameId : <string>
        mode : selectSingle
    }
    sessionId : <string>
}

Running test: testDirectoryPickerAPI
Page.fileChooserOpened {
    method : Page.fileChooserOpened
    params : {
        frameId : <string>
        mode : selectSingle
    }
    sessionId : <string>
}
